new schema with users and authentication

+++ internal/service/account.go
@@ -2,7 +2,7 @@ package service
 
 import (
 	"context"
-	"strconv"
+	"fmt"
 	"time"
 
 	"github.com/Viquad/crud-app/internal/domain"
@@ -10,8 +10,8 @@ import (
 	"github.com/sirupsen/logrus"
 )
 
-var cache_salt = "_account"
-var listId int64 = 0
+const cache_key_template = "user[%d]/account[%d]"
+const listId int64 = 0
 
 type AccountService struct {
 	repo  domain.AccountRepository
@@ -19,34 +19,45 @@ type AccountService struct {
 	ttl   time.Duration
 }
 
-func NewAccountService(repo domain.AccountRepository, cache cache.Cache, ttl time.Duration) *AccountService {
+func NewAccountService(repo Repositories, cache cache.Cache, ttl time.Duration) *AccountService {
 	return &AccountService{
-		repo:  repo,
+		repo:  repo.GetAccountRepository(),
 		cache: cache,
 		ttl:   ttl,
 	}
 }
 
-func (s *AccountService) Create(ctx context.Context, accountInput domain.Account) (*domain.Account, error) {
-	account, err := s.repo.Create(ctx, accountInput)
+func (s *AccountService) Create(ctx context.Context, input domain.AccountCreateInput) (*domain.Account, error) {
+	userId, ok := ctx.Value(domain.UserIdKey).(int64)
+	if !ok {
+		return nil, domain.ErrInvalidId
+	}
+
+	account, err := s.repo.Create(ctx, input)
 	if err == nil {
-		s.cache.Set(idToString(account.Id), account, s.ttl)
+		s.cache.Set(cacheKey(userId, account.Id), account, s.ttl)
+		s.cache.Delete(cacheKey(userId, listId))
 	}
 
 	return account, err
 }
 
 func (s *AccountService) GetById(ctx context.Context, id int64) (account *domain.Account, err error) {
-	i, err := s.cache.Get(idToString(id))
-	var ok bool
+	userId, ok := ctx.Value(domain.UserIdKey).(int64)
+	if !ok {
+		return nil, domain.ErrInvalidId
+	}
+
+	var cached bool
+	i, err := s.cache.Get(cacheKey(userId, id))
 	if err == nil {
 		logrus.WithFields(logrus.Fields{
 			"context": "AccountService.GetById()",
 		}).Debug("Get account from cache")
-		account, ok = i.(*domain.Account)
+		account, cached = i.(*domain.Account)
 	}
 
-	if !ok {
+	if !cached {
 		logrus.WithFields(logrus.Fields{
 			"context": "AccountService.GetById()",
 		}).Debug("Get account from repo")
@@ -54,23 +65,28 @@ func (s *AccountService) GetById(ctx context.Context, id int64) (account *domain
 	}
 
 	if err == nil {
-		s.cache.Set(idToString(id), account, s.ttl)
+		s.cache.Set(cacheKey(userId, id), account, s.ttl)
 	}
 
 	return account, err
 }
 
 func (s *AccountService) List(ctx context.Context) (accounts []domain.Account, err error) {
-	i, err := s.cache.Get(idToString(listId))
-	var ok bool
+	userId, ok := ctx.Value(domain.UserIdKey).(int64)
+	if !ok {
+		return nil, domain.ErrInvalidId
+	}
+
+	var cached bool
+	i, err := s.cache.Get(cacheKey(userId, listId))
 	if err == nil {
 		logrus.WithFields(logrus.Fields{
 			"context": "AccountService.List()",
 		}).Debug("Get accounts from cache")
-		accounts, ok = i.([]domain.Account)
+		accounts, cached = i.([]domain.Account)
 	}
 
-	if !ok {
+	if !cached {
 		logrus.WithFields(logrus.Fields{
 			"context": "AccountService.List()",
 		}).Debug("Get accounts from repo")
@@ -78,30 +94,41 @@ func (s *AccountService) List(ctx context.Context) (accounts []domain.Account, e
 	}
 
 	if err == nil {
-		s.cache.Set(idToString(listId), accounts, s.ttl)
+		s.cache.Set(cacheKey(userId, listId), accounts, s.ttl)
 	}
 
 	return accounts, err
 }
 
 func (s *AccountService) UpdateById(ctx context.Context, id int64, inp domain.AccountUpdateInput) (*domain.Account, error) {
+	userId, ok := ctx.Value(domain.UserIdKey).(int64)
+	if !ok {
+		return nil, domain.ErrInvalidId
+	}
+
 	account, err := s.repo.UpdateById(ctx, id, inp)
 	if err == nil {
-		s.cache.Set(idToString(account.Id), account, s.ttl)
+		s.cache.Set(cacheKey(userId, account.Id), account, s.ttl)
+		s.cache.Delete(cacheKey(userId, listId))
 	}
 
 	return s.repo.UpdateById(ctx, id, inp)
 }
 
 func (s *AccountService) DeleteById(ctx context.Context, id int64) error {
+	userId, ok := ctx.Value(domain.UserIdKey).(int64)
+	if !ok {
+		return domain.ErrInvalidId
+	}
+
 	err := s.repo.DeleteById(ctx, id)
 	if err == nil {
-		s.cache.Delete(idToString(id))
+		s.cache.Delete(cacheKey(userId, id))
 	}
 
 	return err
 }
 
-func idToString(id int64) string {
-	return strconv.FormatInt(id, 10) + cache_salt
+func cacheKey(user_id, id int64) string {
+	return fmt.Sprintf(cache_key_template, user_id, id)
 }
